Move BeaconStateImpl to sub-package
