+++ consensus/src/main/java/org/ethereum/beacon/consensus/transition/InitialStateTransition.java
@@ -12,8 +12,8 @@ import org.ethereum.beacon.consensus.StateTransition;
 import org.ethereum.beacon.consensus.state.ValidatorRegistryUpdater;
 import org.ethereum.beacon.core.BeaconBlock;
 import org.ethereum.beacon.core.BeaconState;
-import org.ethereum.beacon.core.BeaconStateImpl;
-import org.ethereum.beacon.core.BeaconStateImpl.Builder;
+import org.ethereum.beacon.core.state.BeaconStateImpl;
+import org.ethereum.beacon.core.state.BeaconStateImpl.Builder;
 import org.ethereum.beacon.core.operations.Deposit;
 import org.ethereum.beacon.core.spec.ChainSpec;
 import org.ethereum.beacon.core.state.*;
