+++ consensus/src/test/java/org/ethereum/beacon/consensus/transition/InitialStateTransitionTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Random;
 import org.ethereum.beacon.core.BeaconBlocks;
 import org.ethereum.beacon.core.BeaconState;
-import org.ethereum.beacon.core.BeaconStateImpl;
+import org.ethereum.beacon.core.state.BeaconStateImpl;
 import org.ethereum.beacon.core.operations.Deposit;
 import org.ethereum.beacon.core.spec.ChainSpec;
 import org.ethereum.beacon.pow.DepositContract;
