+++ core/src/main/java/org/ethereum/beacon/core/state/BeaconStateImpl.java
@@ -1,4 +1,4 @@
-package org.ethereum.beacon.core;
+package org.ethereum.beacon.core.state;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.unmodifiableList;
@@ -8,14 +8,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.ethereum.beacon.core.BeaconState;
 import org.ethereum.beacon.core.operations.CustodyChallenge;
-import org.ethereum.beacon.core.state.DepositRootVote;
-import org.ethereum.beacon.core.state.CrosslinkRecord;
-import org.ethereum.beacon.core.state.ForkData;
-import org.ethereum.beacon.core.state.PendingAttestationRecord;
-import org.ethereum.beacon.core.state.ShardCommittee;
-import org.ethereum.beacon.core.state.ShardCommittees;
-import org.ethereum.beacon.core.state.ValidatorRecord;
 import tech.pegasys.artemis.ethereum.core.Hash32;
 import tech.pegasys.artemis.util.uint.UInt64;
